Create library.java
This is a Java program that models a library using inheritance.
There is a parent class called readingMaterial which has two attributes, title and price, and a constructor that takes these two attributes as parameters and prints them out.
There are three child classes that inherit from readingMaterial: book, magazine, and cd. Each child class has an additional attribute (isbn, date, and time, respectively), and a constructor that takes the parent class attributes as well as its own attribute as parameters and prints them out.
Finally, there is a Library class that contains the main method. This method creates one object of each child class and prints out their attributes using the constructors of the parent and child classes.