Exception handling Program
This Java program is an example of exception handling using user-defined exceptions.
The program defines a class check with instance variables name and roll_call, and methods accept(), Checkname(), and Checkrollnum(). The accept() method prompts the user to enter their name and roll number, and assigns the input to the corresponding instance variables.
The Checkname() method checks if the name instance variable contains only alphabets. If the name contains any other characters, it throws a custom exception invalidname with an error message. The Checkrollnum() method checks if the roll_call instance variable is between 0 and 23. If it is not, it throws a custom exception invalidroll with an error message.
The invalidname and invalidroll classes are defined as nested classes inside the check class. Both of these classes extend the built-in Exception class and have a constructor that takes a string parameter for the error message.
The run class contains the main() method, which creates an instance of the check class, calls the accept() method to get user input, and then calls the Checkname() and Checkrollnum() methods to perform the respective checks. If either of the methods throws an exception, the error message is printed to the console.